requirements.txt
tensorflow
scikit-learn
numpy
pandas
ipykernel
matplotlib
Pillow

fastapi[standard]
uvicorn
jinja2